Llamando a la policy de post dentro de las funciones update y destroy

+++ api-rest/app/Http/Controllers/Api/PostController.php
@@ -9,6 +9,7 @@ use App\Models\Tag;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controllers\HasMiddleware;
 use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 
 class PostController extends Controller implements HasMiddleware
@@ -57,7 +58,10 @@ class PostController extends Controller implements HasMiddleware
 
     public function tags(Request $request, Post $post)
     {
-        $data = $request->validate([
+        // Para permitir policies declaradá en el archivo PostPolicy
+        Gate::authorize('author', $post);
+
+        $request->validate([
             'tags' => 'required|array|min:1'
         ]);
 
@@ -89,7 +93,10 @@ class PostController extends Controller implements HasMiddleware
      * Update the specified resource in storage.
      */
     public function update(Request $request, Post $post)
-    {
+    {   
+        // Para permitir policies declaradá en el archivo PostPolicy
+        Gate::authorize('author', $post);
+
         $data = $request->validate([
             'title' => 'required',
             'slug' => 'required|unique:posts,slug,' . $post->id,
@@ -116,6 +123,9 @@ class PostController extends Controller implements HasMiddleware
      */
     public function destroy(Post $post)
     {
+        // Para permitir policies declaradá en el archivo PostPolicy
+        Gate::authorize('author', $post);
+
         $post->delete();
         return response()->json(null, 204);
     }
